template/requirements.txt
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-reflection==1.60.0
isort==5.13.2
loguru==0.7.2
protobuf==4.25.1
pycodestyle==2.11.1
pyflakes==3.1.0
ujson==5.9.0
